PR26 - Melhorias e padronizações.
