+++ src/InOut.Service/Services/Interfaces/IUserAccountService.cs
@@ -1,0 +1,9 @@
+﻿using InOut.Domain.Models.User;
+
+namespace InOut.Service.Services.Interfaces
+{
+    public interface IUserAccountService
+    {
+        Task<UserAccountModel> UpdateUserAccountInfo(UserAccountModel userAccountModel);
+    }
+}
