+++ src/InOut.Service/Services/Interfaces/IUserService.cs
@@ -2,6 +2,6 @@
 {
     public interface IUserService
     {
-        Task<bool> ExistsByCpfCnpj(string cpfCnpj);
+        
     }
 }
